+++ mycodo/inputs/dht11.py
@@ -20,7 +20,7 @@ INPUT_INFORMATION = {
     'options_disabled': ['interface'],
 
     'dependencies_module': [
-        ('internal', 'pigpio', 'pigpio')
+        ('internal', 'pip-exists pigpio', 'pigpio')
     ],
     'interfaces': ['GPIO']
 }